Get rid of the dependency on Treex
This commit adds three files that are part of Treex and DeriMor
requires them for proper operation. Treex::Core::Log was simply
copied from Treex Git as of Fri Oct 20 22:56:14 2017 +0200, commit
55729493216cc59938d78a68578e4d0a1100da93. Treex::Tool::Lexicon::CS
and Treex::Tool::Lexicon::Generation::CS were copied and edited to
exclude all unnecessary stuff that induces dependencies on the rest
of Treex.

This solution is, of course, extremely hacky. But it should work
for the time being, the Perl API is being discontinued anyway.