Change the Perl API DB version to reflect reality
The Perl build system used to produce version called 1.5b1, because
it was supposed to be the input for building 1.5. That never happened
and since we are integrating both build systems together now, we
should rename the version to match its position in the joint pipeline,
which is (the first half of) 1.3.