Allow building the DB with a single command
The build process works in steps which roughly correspond to individual
versions – new annotations are usually added to a new directory and
accumulated there until it's time to release. The old Perl pipeline
supported rebuilding the whole thing from scratch with a single `make`
invocation, but the new Python version doesn't. This commit adds a new
Makefile that extends this capability to the new versions not done in
Perl.

The Makefile is recursive, which is evil. It doesn't detect changes in
subprojects, so you if you make changes to the individual versions,
you'll have to `make clean` from time to time.