Fix lexeme_exists() for empty IDs.
Previously, lexeme_exists checked the pretty_id of the lexeme
unconditionally, but if the user added a node with pretty_id
set to None, the ID was formatted to "" (the empty string) and
equality tests failed.

Change lexeme_exists so that if the ID is empty, it is ignored.
This is useful in the case the user doesn't care for IDs. It is
arguably a hack, though.