Get rid of the underscore in print_lexemes.py
As stated in Issue #14, we want to be able to call modules using
a single name. Right now, the user has to specify both a file name
and a class name. Reducing it to just one of them is desirable.

We do that by using the class name only and deriving the file name
from it by lowercasing it. This approach is incompatible with
underscores in file names.

Therefore, rename print_lexemes.py to printlexemes.py to match its
class name of PrintLexemes.