Optionally ignore extra POS markers in stats
DeriNet 1.5 introduced a way of marking selected lexemes as either
nonderived or composite by adding an extra character to the POS tag.
This breaks the older API, which is especially important for the
statistics calculation code, as it is not superseded by a new tool
yet. Make the tool optionally ignore the new flags.