Quick fix for the homonym confusion in get_id()
When searching for lexemes by full Nodes, the method get_id extracted
the lemma and used that as a query. This caused mismatches for
homonyms. This code is a quick fix/workaround for that. Fixes #16.

A full solution would require transitioning to using IDs only.
That requires some restructuring of the code. Right now, I will not
introduce breaking changes into the code.