Add a method for getting a non-homonymous lexeme
Currently, there are two ways of retrieving lexemes from the database:
get_ids, which returns a list of all matches, and get_id, which returns
any one of them. But there are places where we want to retrieve
a single specific ID (like get_id) but without ignoring homonymy
(like get_ids) – e.g. when retrieving a parent or a child of a lexeme
based on a textual representation.

The new method, get_unique_id, does exactly this. If the lexeme
specification is ambiguous, it raises LexemeAmbiguousError.

Closes #18.