Rewrite lexeme_exists to have a common signature
Most other functions take either a Node, or an ID, or a
lemma-pos-techlemma triplet as arguments. The function lexeme_exists
didn't. Convert it to the common argument format.

This causes a functional change in the handling of pretty_ids, which
were previously compared but now they aren't. Instead, we compare
lex_ids now. I believe this is safe to do, thanks to the planned
obsolescence of pretty_ids and the requirement that all Nodes be
created via the concrete DeriNet instance's methods.