Give add_lexeme() non-opaque checkable args
Previously, add_lexeme() accepted a complete namedtuple Node as an
argument. This created problems, because some fields of the namedtuple
were checked, some were unchecked and some were flat out ignored and
recreated, as they are considered internal to the implementation.

The new implementation takes the innards of the Node as arguments and
doesn't provide a way of specifying the args which were previously
ignored or forbidden, thus making the interface more intuitive and
safer.

This also means that all Node creation is under our control, the user
is never forced to create a Node on their own. Thus, this closes #17.