Hack-fix scenario processing and add a test
The scenario was unable to load modules with shortened path (e.g. only
`cs.Module` instead of the full `derinet.modules.cs.Module`). Fix this
by anchoring the search root to `derinet.modules`. This disables the
full path loading, but that is not so useful anyway (for now).

Also, add a Makefile target that tests the scenario support by loading
and saving a database.