Add relations to all incident lexemes
There is only one main source and one main target and the other ones
are auxiliary. But there are no first-tier and second-tier relations,
those are all the same. Therefore, even for compounding, add the
relation itself even to the secondary sources, which don't have the
main target listed among their children.