Store the lexemes of V1 files in load order
The V1 files have IDs, which should ideally stay the same after
processing the file with a pipeline. The new API doesn't explicitly
support the IDs, but if they are consecutive and ordered, things will
appear to work. This change ensures that we actually keep them as
intact as possible.

Also, it changes the code for compatibility with implementations which
don't have ordered dicts as the default.