When saving V2, go over the lexemes in two passes
This change is necessary, because while in-tree lexemes may only ever
be processed in tree order, secondary relations do exist and those may
be processed out of order. We therefore need to generate IDs of all
trees in the database before we start printing the lexemes, because any
lexeme may refer to an ID of any other lexeme.