Add the root morphemes to the database properly
Until now, the root morphemes were only annotated in the misc section
and the module didn't use the official methods for adding them as
segmentation. Fix that.

For now, this is a bit of a hack, since the annotation is added to both
the misc section and to the segmentation section. In the future, it
would be good if we got rid of the data in misc.