Add treechecking module based on enumerated words
In Czech, the spelling of i/y after certain consonants depends on
whether the word is on a list of enumerated words or is derived from
one. This derivation criterion can be used for checking the correctness
of our trees: If we find a word with [bflmpsvz][yý] in the lemma, it
must be enumerated or descended from one. If it is not, we are missing
a link.