Escape allomorphs before building regex
The allomorphs contain special characters, if the annotators didn't
fill in everything. They use e.g. `%?` as a sign that the root morph
is unclear. That is, in effect, the same thing as an empty allomorph,
because it causes misparsing when the regex is used.

Escape the allomorphs properly, so that the question marks and other
special characters don't mess things up.