Move CorpusCount to corpus_stats.absolute_count
The `CorpusCount` key identifies a single scalar value, but we want to
include more types of values. Also, the capitalization is not proper
for JSON, although it is proper for the rest of DeriNet keys.

This commit introduces a hierarchy under the `corpus_stats` key,
currently with only `absolute_count` in there. But we want to add more
scores in the future.