Implement getting morphs from CELEX morphemes
The CELEX databases contain annotation of morphological segmentation,
but they only list the (disambiguated) morphemes, not the exact morph
substrings from the lemma they correspond to.

But sometimes, the allomorph chosen as the representation of the
morpheme happens to be the one in the lemma as well. In these cases,
we can delimit the morphs easily. Implement that as a buildsystem
module.