Segment even lemmas with some limited allomorphy
Sometimes, it is possible to map even morphemes with some phonological
or other changes to morphs. It can be easily done if only one morpheme
has allomorphy present – the other morphemes are delimited first and
the allomorphic one corresponds to the unmapped remainder of the lemma.

This code has some limitations, see the FIXME notes there.