Add the possibility to ignore POS differences
When measuring precision and recall against data with no POS tags,
or when measuring the stats on a database with no POS tags, it is
beneficial to ignore the tags when finding which lexemes to compare
againt which ones.

This is not used in the Czech DeriNet, but it is useful for other
languages.