Add a module for copying aspect from old versions
The new MorfFlex, which we now use, lacks required annotations of
verbal aspect in the technical suffixes of the lemmas. But for the
most part, we can just copy those from an older version of the
database.

As usual, homonyms complicate the process and some verbs from the
new dictionary are not present in 2.0, but apart from ca. 800 problems,
the rest of the verbs are now annotated again.