Change `_index` to store indices into `_data`
When deleting lexemes, we have to find where in `_data` is the lexeme
stored. Since this information was not available anywhere, we scanned
the whole list, which was slow.

If we change `_index` to hold indices to `_data` instead of references
to the lexemes themselves, we add one layer of indirection to
`get_lexemes`, but turn an O(n) deletion into O(homonym count), which
is usually O(1).