Document the recommender roundtrip test
With version 2.1, we accidentally released data that can't be cleanly
loaded by the API due to a bug in cycle detection. The bug has been
fixed, but a released version can't be taken back. To avoid this in the
future, we should always try to load the released data and verify that
no errors were reported.