Fix main univerbation parent
The univerbation relations were previously always connected to the
last parent as the main one, because it is the one from which the
grammatical categories are derived. However, a better and
linguistically more adequate solution is to choose the parent with
which the lexeme shares its root morpheme(s) as the main one.

Fix this using a similarity heuristic, as there is no hard rule as to
whether the main parent is the first or last one.