Restore process_scenario.py in the orig location
There should be no in-tree scripts left still using process_scenario.py
as a directly-executed script, everything should've been converted to
use it as a module. Nonetheless, external code may still reference the
old location and use the script directly. Restore support for that by
including a symlink from the old location to the new one.