Parse derivational comments with two stars
The two stars mean that the lexeme's lemma should be fully replaced by
the given string to obtain the parent's lemma. These derivations are
often marginal or entirely undesirable (Jan -> Honza), but they seem
somewhat more useful for numerals. We can turn them off again if we
find that they are not worth adding.