Refactor code and output format
Rename variables. Split the files reporting generated derivations into
those where the derivation lexeme was root of a tree so it was directly
conected to the parent root node.
And those where the derivation lexeme is not root so the root of this
derivation is connected to the parent root, there could be some false
derivations if this is the case.