Add file with added indirect derivations
There are ~455 derivations generated, manualy flaged which are probably
incorrecThere are ~455 derivations generated, manualy flaged which are
probably incorrect.
Cca 341 new probably correct derivations (cca 75 % precision).
The node to add as derivation is not root of a tree, so insted we
connect the root of this tree to the parent. The original derivation
(the word that is in the Etym dict) is in this file in column
intermediate node.