Introduce special methods for removing relations
Until now, removing relations was done using
`lexeme.parent_relation.remove_from_lexemes()`. The issue with this
method is twofold:
1. Adding relations is done through a method on the Lexicon object, so
   removal should ideally be similar, and
2. the `remove_from_lexemes()` method cannot access the Lexicon context
   to record *who* did the removal, only that it happened.

Introduce a new set of methods on Lexicon to perform relation removal,
refactor the relation removal in the lexeme deletion method to use them
and document the caveat with deleting relations in a loop.

Code in other modules which remove relations is not refactored yet.