Initialize modules just before running them
Early initialization has the advantage that some errors in the modules
can be discovered early, before the time-consuming loading of the
lexicon and running of other modules take place. However, if we want to
record relation changes, we should preserve the args and kwargs which
the modules get until processing. And if we're doing that anyway, it
makes sense to pass the uninitialized class with the args and init it
just in time.