Implement storing history when adding relations
To allow tracing changes and easy identification of error sources, add
methods for storing the history of relation changes. The history is
stored with the child lexeme and allows recording version of the DB,
the name of the creator module and its args. So far, the code only
records additions, not deletions, and it is not actually activated by
the scenario processor.