Add the lexeme etymology information
Note that due to the way Feats are currently serialized, the etymology
is added as a list, but after saving and reloading the lexicon, gets
turned into a string. This should be fixed before release.

Also note that the data are currently not publicly available. They will
be either committed later, or the filename will be changed to reference
a location on the UFAL network.

References #23