Add machinery for selecting the count file format
The counts from SYNv9 have a different format, using positional
morphological tags instead of just the POS and differentiating lemma
and sublemma, neither of which maps cleanly to our concept of lemmas.
This means we'll need different code to handle v9 and v4. Add a
command-line option to choose between the two, even though only the v4
side is populated right now.