Honor on_err when saving otherrels to V1 format
When there are multiple derivations present in a single lexeme, the old
V1 file format cannot express them and only one will be saved. Usually,
this is a hard error. But when `on_err="continue"` is specified, only
issue a warning, save the main parent of the main relation and continue
saving.