Fix MorfFlex error with possessive ADJ PossGender
Possessive adjectives ending with -ýma are homonymous between masculine
and feminine PossGender (Karlovýma and Lenčinýma). MorfFlex
systematically tags all examples as masculine PossGender. Ignore the
PossGender when encountering an -ýma word with tag AU?P7M--------7, so
that the PossGender is properly preserved for feminine-based adjectives
and not replaced with `?` in the tag mask.