Downcase keys for JSON printing, retitle on load
JSON typically uses lowercased keys, while our API really needs
TitleCased ones. Downcase the commonly present keys on printing and
re-TitleCase them on load. This is impossible to solve generally
(specifically the retitlecasing part), so other keys will stay in
whatever case the user entered when adding the segmentation, but
I think that's fine, actually, as the main issue here is internal
API usage of the enumerated keys, which are handled correctly.