Add setup.py for compatibility with older Python
We still keep compatibility with Python 3.6, at least for the core API
(some modules might not be compatible, although they generally are).
The newest version of setuptools that works with Python 3.6 is however
still too old to be able to process `pyproject.toml`. Add a simple
`setup.py` to allow installing the API using Pip even on such old
versions.