Convert Block's `parse_args()` into classmethod
The `parse_args()` method needs access to the class name in order to
provide it as the module identification to command-line users through
`argparse`.

This is currently achieved through the lexically-scoped `__class__`
variable, but that has the issue that any subclasses which don't
override the `parse_args()` method get the lexically-scoped value
`Block` reported as the module name in their command-line help.

By converting the method from `@staticmethod` to `@classmethod`, we
gain access to the class object, which dynamically reflects the actual
subclass that is being used.