Implement parsing the Loanword key as a boolean
The check for special parsing and serialization of Loanword is done in
common code, which is also used for handling morphemes and relation
feats. This might need fixing in the future by either splitting the
methods into lex-feats/rel-feats/morph specific ones, or by adding
another parameter specifying this.

References #23